Avoid duplication in make_standalone with multiple encoders
This corrects a problem in the recent 6d673f6 (Extend make_standalone
script to allow for more than 1 encoder, 2021-02-12) which could (and
did) result in having duplicated dependencies in the generated files.
While this didn't seem to result in any problems, it is still at best
completely unnecessary.

Keep track of the dependencies we had already seen to prevent this from
happening.